+++ frontend/tailwind.config.js
@@ -1,4 +1,5 @@
 import forms from '@tailwindcss/forms';
+import typography from '@tailwindcss/typography';
 
 /** @type {import('tailwindcss').Config} */
 export default {
@@ -124,5 +125,6 @@ export default {
   },
   plugins: [
     forms,
+    typography,
   ],
 }